+++ src/main/java/wooteco/subway/domain/fare/AgeDiscountPolicy.java
@@ -1,7 +1,5 @@
 package wooteco.subway.domain.fare;
 
-import wooteco.subway.domain.policy.FarePolicy;
-
 public interface AgeDiscountPolicy extends FarePolicy {
 
     @Override
