+++ src/main/java/wooteco/subway/domain/fare/DistancePolicy.java
@@ -1,5 +1,5 @@
 package wooteco.subway.domain.fare;
 
 public interface DistancePolicy {
-    int getFare();
+    int getFare(int distance);
 }
