+++ src/main/java/wooteco/subway/domain/fare/FarePolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.policy;
+package wooteco.subway.domain.fare;
 
 public interface FarePolicy {
     double calculate(double baseFare);
