+++ src/main/java/wooteco/subway/domain/fare/policy/distance/OverFiftyKMPolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.policy.distance;
+package wooteco.subway.domain.fare.policy.distance;
 
 import wooteco.subway.domain.fare.DistancePolicy;
 
@@ -7,14 +7,8 @@ public class OverFiftyKMPolicy implements DistancePolicy {
     private static final int OVER_DISTANCE = 50;
     private static final double OVER_RATE = 8;
 
-    private final int distance;
-
-    public OverFiftyKMPolicy(int distance) {
-        this.distance = distance;
-    }
-
     @Override
-    public int getFare() {
+    public int getFare(int distance) {
         return BASE_FEE + (int) (Math.ceil((distance - OVER_DISTANCE) / OVER_RATE) * 100);
     }
 }
