+++ src/main/java/wooteco/subway/domain/fare/policy/distance/TenToFiftyKMPolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.policy.distance;
+package wooteco.subway.domain.fare.policy.distance;
 
 import wooteco.subway.domain.fare.DistancePolicy;
 
@@ -7,14 +7,8 @@ public class TenToFiftyKMPolicy implements DistancePolicy {
     private static final int OVER_TEN_DISTANCE = 10;
     private static final double OVER_TEN_RATE = 5;
 
-    private final int distance;
-
-    public TenToFiftyKMPolicy(int distance) {
-        this.distance = distance;
-    }
-
     @Override
-    public int getFare() {
+    public int getFare(int distance) {
         return BASE_FEE + (int) (Math.ceil((distance - OVER_TEN_DISTANCE) / OVER_TEN_RATE) * 100);
     }
 }
