+++ src/main/java/wooteco/subway/domain/fare/policy/distance/UnderTenKMPolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.policy.distance;
+package wooteco.subway.domain.fare.policy.distance;
 
 import wooteco.subway.domain.fare.DistancePolicy;
 
@@ -6,7 +6,7 @@ public class UnderTenKMPolicy implements DistancePolicy {
     static final int BASE_FEE = 1250;
 
     @Override
-    public int getFare() {
+    public int getFare(int distance) {
         return BASE_FEE;
     }
 }
